Schemas improvements

+++ src/services/OrionAggregator/schemas/pairConfigSchema.ts
@@ -1,16 +1,16 @@
 import { z } from 'zod';
 
 const pairConfigSchema = z.object({
-  baseAssetPrecision: z.number(),
+  baseAssetPrecision: z.number().int(),
   executableOnBrokersPriceDeviation: z.number().nullable(),
   maxPrice: z.number(),
   maxQty: z.number(),
   minPrice: z.number(),
   minQty: z.number(),
   name: z.string(),
-  pricePrecision: z.number(),
-  qtyPrecision: z.number(),
-  quoteAssetPrecision: z.number(),
+  pricePrecision: z.number().int(),
+  qtyPrecision: z.number().int(),
+  quoteAssetPrecision: z.number().int(),
 });
 
 export default pairConfigSchema;
